% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Addition.R
\name{Addition}
\alias{Addition}
\title{Addition}
\usage{
Addition(x, y)
}
\arguments{
\item{x}{First value}

\item{y}{Second value}
}
\value{
z sum of values
}
\description{
Addition
}
\examples{
#Output = 3
Addition(1,2)
}
