% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Pokemon}
\alias{Pokemon}
\title{A dataset containing a lot of pokemon information}
\format{
\subsection{\code{Pokemon}}{

\describe{
\item{name}{Pokemon name}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/rounakbanik/pokemon/}
}
\usage{
Pokemon
}
\description{
A Kaggle dataset of a serebii scraped dataset
}
\keyword{datasets}
